record created archive nodes before propagating copy errors
A destination node created by the copier was only added to the cleanup
manifest once its entire entry succeeded. When a later step failed --
io.Copy returning ENOSPC/EIO, a failing Close, a fresh directory that
could not be reopened -- the helper returned an empty entry and the
level bailed out before recording it. Cleanup then saw a node it had no
manifest entry for, read it as unexpected, and refused to remove the
.af-copy-* staging tree, so every retried archive stranded another
partial copy in the destination parent.

Each helper now identifies the node it just created and returns that
entry alongside the error; copyDirectoryLevel records any named entry
before propagating. A named entry means "this process created this node
and knows its identity", so cleanup still only ever removes nodes whose
identity it recorded -- taking the identity immediately after creation
narrows that window rather than widening it. copyDirectoryEntry also now
verifies the reopened directory is the one it created.

Splits the copier out of worktree_archive.go into worktree_copy_tree.go
alongside worktree_copy_cleanup.go to stay under the file-length limit.
