+++ session/git/worktree_copy_cleanup_test.go
@@ -1,6 +1,7 @@
 package git
 
 import (
+	"fmt"
 	"os"
 	"path/filepath"
 	"strings"
@@ -68,6 +69,106 @@ func TestCopyTree_RejectsDestinationSymlinkReplacement(t *testing.T) {
 	t.Cleanup(func() { copyTreeAfterSymlinkCreate = originalHook })
 
 	require.Error(t, copyTree(src, dest), "a replacement symlink target must not enter the copy manifest")
+}
+
+// TestCopyTree_PostCreateFailureCleansPartialDestination pins the manifest
+// contract: a destination node is recorded the moment this process creates it
+// and learns its identity, not once the whole entry succeeds. A failure after
+// creation — ENOSPC mid-copy, a failing Close, a directory that cannot be
+// reopened — must leave a removable partial tree rather than one cleanup
+// refuses as unexpected.
+func TestCopyTree_PostCreateFailureCleansPartialDestination(t *testing.T) {
+	for _, testCase := range []struct {
+		name    string
+		build   func(t *testing.T, src string)
+		created string
+	}{
+		{
+			name: "regular file",
+			build: func(t *testing.T, src string) {
+				require.NoError(t, os.WriteFile(filepath.Join(src, "a-file"), []byte("original"), 0644))
+			},
+			created: "a-file",
+		},
+		{
+			name: "directory",
+			build: func(t *testing.T, src string) {
+				require.NoError(t, os.Mkdir(filepath.Join(src, "a-dir"), 0755))
+			},
+			created: "a-dir",
+		},
+	} {
+		t.Run(testCase.name, func(t *testing.T) {
+			src := filepath.Join(t.TempDir(), "src")
+			require.NoError(t, os.Mkdir(src, 0755))
+			testCase.build(t, src)
+			dest := filepath.Join(t.TempDir(), "dest")
+
+			originalHook := copyTreeAfterDestCreate
+			copyTreeAfterDestCreate = func(path string) error {
+				if !strings.HasSuffix(path, testCase.created) {
+					return nil
+				}
+				return fmt.Errorf("forced failure after destination create: %w", syscall.ENOSPC)
+			}
+			t.Cleanup(func() { copyTreeAfterDestCreate = originalHook })
+
+			err := copyTree(src, dest)
+			require.ErrorIs(t, err, syscall.ENOSPC)
+			assert.NotContains(t, err.Error(), "refusing to remove",
+				"a node this process created must stay removable, not read as unexpected")
+			assert.NoDirExists(t, dest, "the partial destination tree must not leak")
+		})
+	}
+}
+
+// TestCopyTree_SymlinkPostCreateFailureCleansPartialDestination covers the same
+// contract on the symlink path, whose creation seam predates the manifest entry.
+func TestCopyTree_SymlinkPostCreateFailureCleansPartialDestination(t *testing.T) {
+	src := filepath.Join(t.TempDir(), "src")
+	require.NoError(t, os.Mkdir(src, 0755))
+	require.NoError(t, os.Symlink("source-target", filepath.Join(src, "link")))
+	dest := filepath.Join(t.TempDir(), "dest")
+
+	originalHook := copyTreeAfterSymlinkCreate
+	copyTreeAfterSymlinkCreate = func(string) error {
+		return fmt.Errorf("forced failure after symlink create: %w", syscall.EIO)
+	}
+	t.Cleanup(func() { copyTreeAfterSymlinkCreate = originalHook })
+
+	err := copyTree(src, dest)
+	require.ErrorIs(t, err, syscall.EIO)
+	assert.NotContains(t, err.Error(), "refusing to remove")
+	assert.NoDirExists(t, dest, "the partial destination tree must not leak")
+}
+
+// TestMoveDirCrossDevice_PostCreateFailureCleansPrivateStaging is the user-facing
+// consequence: without the created node in the manifest, every failed archive
+// attempt strands another .af-copy-* tree in the destination parent.
+func TestMoveDirCrossDevice_PostCreateFailureCleansPrivateStaging(t *testing.T) {
+	src := filepath.Join(t.TempDir(), "src")
+	require.NoError(t, os.Mkdir(src, 0755))
+	require.NoError(t, os.WriteFile(filepath.Join(src, "a-file"), []byte("original"), 0644))
+	destinationParent := t.TempDir()
+
+	originalRename := renamePath
+	renamePath = func(_, _ string) error { return syscall.EXDEV }
+	t.Cleanup(func() { renamePath = originalRename })
+	originalHook := copyTreeAfterDestCreate
+	copyTreeAfterDestCreate = func(path string) error {
+		if !strings.HasSuffix(path, "a-file") {
+			return nil
+		}
+		return fmt.Errorf("forced failure after destination create: %w", syscall.ENOSPC)
+	}
+	t.Cleanup(func() { copyTreeAfterDestCreate = originalHook })
+
+	err := moveDirCrossDevice(src, filepath.Join(destinationParent, "dest"))
+	require.ErrorIs(t, err, syscall.ENOSPC)
+	entries, readErr := os.ReadDir(destinationParent)
+	require.NoError(t, readErr)
+	assert.Empty(t, entries, "a failed copy must not strand a private staging tree")
+	assert.FileExists(t, filepath.Join(src, "a-file"), "a failed move must leave the source intact")
 }
 
 func TestCopyTree_RejectsExcessiveDepth(t *testing.T) {
